Extracted logic for deleting the Monitor in separate service class.
